pyproject.toml
[project]
name = "training-backend"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "sqlalchemy>=2.0",
    "psycopg[binary]",
    "pydantic[email]",
    "alembic",
    "python-jose[cryptography]"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 88